paralegal_flowistry::mir::placeinfo
Struct PlaceInfo

pub struct PlaceInfo<'tcx> {
    pub tcx: TyCtxt<'tcx>,
    pub body: &'tcx Body<'tcx>,
    pub def_id: DefId,
    aliases: Aliases<'tcx>,
    normalized_cache: CopyCache<Place<'tcx>, Place<'tcx>>,
    aliases_cache: Cache<Place<'tcx>, PlaceSet<'tcx>>,
    conflicts_cache: Cache<Place<'tcx>, PlaceSet<'tcx>>,
    reachable_cache: Cache<(Place<'tcx>, Mutability), PlaceSet<'tcx>>,
}

Utilities for analyzing places: children, aliases, etc.

Fields
tcx: TyCtxt<'tcx>body: &'tcx Body<'tcx>def_id: DefIdaliases: Aliases<'tcx>normalized_cache: CopyCache<Place<'tcx>, Place<'tcx>>aliases_cache: Cache<Place<'tcx>, PlaceSet<'tcx>>conflicts_cache: Cache<Place<'tcx>, PlaceSet<'tcx>>reachable_cache: Cache<(Place<'tcx>, Mutability), PlaceSet<'tcx>>
Implementations

impl<'tcx> PlaceInfo<'tcx>

pub fn build<'a>( tcx: TyCtxt<'tcx>, def_id: DefId, input: impl FlowistryInput<'tcx, 'a>, ) -> Self

Computes all the metadata about places used within the infoflow analysis.

pub fn build_from_input_facts<'a>( tcx: TyCtxt<'tcx>, def_id: DefId, input: impl FlowistryInput<'tcx, 'a>, ) -> Self

Computes all the metadata about places used within the infoflow analysis.

pub fn body(&self) -> &'tcx Body<'tcx>

pub fn normalize(&self, place: Place<'tcx>) -> Place<'tcx>

Normalizes a place via PlaceExt::normalize (cached).

See the PlaceExt documentation for details on how normalization works.

pub fn aliases(&self, place: Place<'tcx>) -> &PlaceSet<'tcx>

Computes the aliases of a place (cached).

For example, if x = &y, then *x aliases y. Note that an alias is NOT guaranteed to be of the same type as place!

pub fn children(&self, place: Place<'tcx>) -> PlaceSet<'tcx>

Returns all reachable fields of place without going through references.

For example, if x = (0, 1) then children(x) = {x, x.0, x.1}.

pub fn conflicts(&self, place: Place<'tcx>) -> &PlaceSet<'tcx>

Returns all places that directly conflict with place, i.e. that a mutation to place would also be a mutation to the conflicting place.

For example, if x = ((0, 1), 2) then conflicts(x.0) = {x, x.0, x.0.0, x.0.1}, but not x.1.

For indirect places, this function follows conflicting parents up until a reference point. So if x = (0, &(box 1, 2)) then conflicts(((x.1).0)) = {((x.1).0), *(x.1).0, *(x.1)}

pub fn reachable_values( &self, place: Place<'tcx>, mutability: Mutability, ) -> &PlaceSet<'tcx>

Returns all direct places that are reachable from place and can be used at the provided level of [Mutability] (cached).

For example, if x = 0 and y = (0, &x), then reachable_values(y, Mutability::Not) is {y, x}. With Mutability::Mut, then the output is {y} (no x).

fn collect_loans(&self, ty: Ty<'tcx>, mutability: Mutability) -> PlaceSet<'tcx>